songcrud/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
distlib==0.3.6
Django==4.1.2
django-filter==22.1
djangorestframework==3.14.0
filelock==3.8.0
Flask==2.2.2
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
numpy==1.23.4
pandas==1.5.1
platformdirs==2.5.2
PyJWT==2.6.0
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
twilio==7.15.0
tzdata==2022.4
urllib3==1.26.12
virtualenv==20.16.5
Werkzeug==2.2.2